requirements.txt
python-consul>=1.10
